refactor: AI tools in different files
