+++ ex_app/lib/all_tools/context_chat.py
@@ -1,0 +1,35 @@
+# SPDX-FileCopyrightText: 2025 Nextcloud GmbH and Nextcloud contributors
+# SPDX-License-Identifier: AGPL-3.0-or-later
+from langchain_core.tools import tool
+from nc_py_api import Nextcloud
+
+from ex_app.lib.all_tools.lib.task_processing import run_task
+from ex_app.lib.all_tools.lib.decorator import safe_tool
+
+
+def get_tools(nc: Nextcloud):
+
+	@tool
+	@safe_tool
+	def ask_context_chat(question: str) -> str:
+		"""
+		Ask the context chat oracle, which knows all of the user's documents, a question about them
+		:param question: The question to ask
+		:return: the answer from context chat
+		"""
+
+		task_input = {
+			'prompt': question,
+			'scopeType': 'none',
+			'scopeList': [],
+			'scopeListMeta': '',
+		}
+		task_output = run_task(nc,  "context_chat:context_chat", task_input).output
+		return task_output['output']
+
+	return [
+		ask_context_chat,
+	]
+
+def get_category_name():
+	return "Context Chat"
